Add macroConnections package and move related classes
