+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/models/macroconnections/MacroConnect.java
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.dynawaltz.models;
+package com.powsybl.dynawaltz.models.macroconnections;
 
 import javax.xml.stream.XMLStreamException;
 import javax.xml.stream.XMLStreamWriter;
