+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/models/macroconnections/MacroConnectAttribute.java
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.dynawaltz.models;
+package com.powsybl.dynawaltz.models.macroconnections;
 
 /**
  * @author Laurent Issertial <laurent.issertial at rte-france.com>
