+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/models/macroconnections/MacroConnector.java
@@ -4,7 +4,10 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dynawaltz.models;
+package com.powsybl.dynawaltz.models.macroconnections;
+
+import com.powsybl.dynawaltz.models.Side;
+import com.powsybl.dynawaltz.models.VarConnection;
 
 import javax.xml.stream.XMLStreamException;
 import javax.xml.stream.XMLStreamWriter;
